created so that the analizator data is ManyToOne and it can hold only one refference to the patient. THe patient is OneToMany so it will hold and array of analizator datas

+++ back-End/src/main/java/studiobox/jobs/hztm_patient_manager/repositorys/PatientDataRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface PatientDataRepository extends JpaRepository<PatientData, Long> {
-    Optional<PatientData> findByOib(int oib);
+    Optional<PatientData> findByOib(Long oib);
 }
